# [PERF] stock: optimize _compute_rules