Originally, we compute rules for stock orderpoints based on product_id through `product_id._get_rules_from_location`, however, this function only relies on `product_id.route_ids` or
`product_id.categ_id.total_route_ids` which usually don't change a lot per product.

This PR leverages this idea that we can simply extend the caching key to include the routes_ids and categ_id.total_route_ids of products so we don't have to calculate the same value twice. For products without route_ids, it simply falls back to the original cache key.

Benchmarks

|Num. orderpoints| Before PR | After PR |
|---------------------|---------------|--------------|
|5636| 56.2  s| 39.5 s|

opw-4649249


---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr