@robodoo 2025-03-18
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/202365.png)](https://mergebot.odoo.com/odoo/odoo/pull/202365)