@MohamedKasem99 2025-03-24
@svs-odoo Please review this whenever you have the time